azure-mi-in-py-sdk/pyproject.toml
[tool.poetry]
name = "azure-mi-in-py-sdk"
version = "0.0.1"
description = ""
authors = ["Alex Ott <alexey.ott@databricks.com>"]
license = "Databricks License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
databricks-sdk = "^0.44.1"
azure-identity = "^1.21.0"

[tool.poetry.group.test.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
